created LAN communication inside Equipment Class
